ADD: MSoE pose estimator by using a Objective Function interface.

+++ cpp/src/objectiveFunctions/ObjectiveFunction.hpp
@@ -9,7 +9,7 @@ class ObjectiveFunction
 {
     public:
         virtual void setPointCloud(const std::vector<Eigen::Vector4d> &pointCloud) = 0;
-        virtual std::vector<double> calculateEvidence(const Eigen::MatrixXd &hypotheses) = 0;
+        virtual std::vector<int> calculateEvidence(const Eigen::MatrixXd &hypotheses) = 0;
         virtual ~ObjectiveFunction() = default;
 };
